+++ backends/carp_webservices/example/lib/homepage.dart
@@ -22,8 +22,12 @@ class _HomePageState extends State<HomePage> {
               color: Colors.blue, borderRadius: BorderRadius.circular(10)),
           child: FlatButton(
             onPressed: () {
-              Navigator.push(context,
-                  MaterialPageRoute(builder: (_) => CarpAuthenticationForm()));
+              Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (_) => CarpAuthenticationForm(
+                            username: 'sys1admin1DK@cachet.dk',
+                          )));
             },
             child: Text(
               'LOGIN',