+++ backends/carp_webservices/test/carp_deployment_service_test.dart
@@ -33,9 +33,9 @@ void main() {
       oauth: OAuthEndPoint(clientID: clientID, clientSecret: clientSecret),
     );
 
-    CarpService.configure(app);
+    CarpService().configure(app);
 
-    user = await CarpService.instance.authenticate(
+    user = await CarpService().authenticate(
       username: username,
       password: password,
     );
@@ -47,25 +47,25 @@ void main() {
 
   group("Deployment", () {
     test('- authentication', () async {
-      print('CarpService : ${CarpService.instance.app}');
+      print('CarpService : ${CarpService().app}');
       print(" - signed in as: $user");
     }, skip: false);
 
     test('- device ID', () async {
-      String id = CarpService.instance.deployment().registeredDeviceId;
+      String id = CarpService().deployment().registeredDeviceId;
       print('Registered Device ID : $id');
     }, skip: false);
 
     test('- get invitations for this account (user)', () async {
       List<ActiveParticipationInvitation> invitations =
-          await CarpService.instance.invitations();
+          await CarpService().invitations();
       invitations.forEach((invitation) => print(invitation));
       //assert(invitations.length > 0);
     }, skip: false);
 
     test('- get deployment status', () async {
       StudyDeploymentStatus status =
-          await CarpService.instance.deployment().getStatus();
+          await CarpService().deployment().getStatus();
       print(_encode(status.toJson()));
       print(status);
       print(status.masterDeviceStatus.device);
@@ -73,7 +73,7 @@ void main() {
     }, skip: false);
 
     test('- register device', () async {
-      DeploymentReference reference = CarpService.instance.deployment();
+      DeploymentReference reference = CarpService().deployment();
       StudyDeploymentStatus status = await reference.getStatus();
       print(status);
       expect(status.masterDeviceStatus.device, isNotNull);
@@ -85,7 +85,7 @@ void main() {
     }, skip: false);
 
     test('- get master device deployment', () async {
-      DeploymentReference reference = CarpService.instance.deployment();
+      DeploymentReference reference = CarpService().deployment();
       StudyDeploymentStatus status = await reference.getStatus();
       print(status);
       expect(status.masterDeviceStatus.device, isNotNull);
@@ -96,7 +96,7 @@ void main() {
     }, skip: false);
 
     test('- deployment success', () async {
-      DeploymentReference reference = CarpService.instance.deployment();
+      DeploymentReference reference = CarpService().deployment();
       StudyDeploymentStatus status_1 = await reference.getStatus();
       MasterDeviceDeployment deployment = await reference.get();
       StudyDeploymentStatus status_2 = await reference.success();
@@ -106,7 +106,7 @@ void main() {
     }, skip: false);
 
     test('- unregister device', () async {
-      DeploymentReference reference = CarpService.instance.deployment();
+      DeploymentReference reference = CarpService().deployment();
       StudyDeploymentStatus status = await reference.getStatus();
       print(status);
       expect(status.masterDeviceStatus.device, isNotNull);